includes_cache: Restructure name cache
Restructure name cache to store the expression type of the name. This
way, the annoter can tell the cache what it's looking for, i.e. a field,
a function or an enumeration constant. The cache can then verify on its
side if it really has what the annoter is looking for.

This is also necessary to avoid field name collisions. The annoter can
now tell the cache side what field from what struct it needs.

Signed-off-by: Jaskaran Singh <jaskaran.singh@collabora.com>